Adding additional error handling
As a result of the change from ui:parts to ui:part for Multiple fields in the Form Language definition, more errors in Form Models are likely to occur. This change is to handle the case where a Multiple without a part defined is an invalid Form Model, and exit the function early.